test(desktop): split execution factory suite
