+++ src/test-kernel/desktop/desktopAgentExecutionTestHarness.mts
@@ -1,0 +1,67 @@
+// Third-party imports
+import { onTestFinished, vi } from 'vitest';
+
+// Local imports - shared schemas
+import type { RunOutcome } from '@shared/schemas';
+
+export type DesktopAgentExecutionModule =
+  typeof import('@desktop/main/desktopAgentExecution');
+
+export type RunExecutionRequest = (
+  request: unknown,
+  options: {
+    openWorkflowOutput(result: {
+      outcome: RunOutcome;
+      outputs: Array<{ absolutePath: string }>;
+    }): Promise<void>;
+    // This window's SessionHandle. The onboarding run-completion test drives a
+    // terminal `result` event through it via `attachRunTrace`.
+    session: {
+      attachRunTrace(
+        trace: {
+          subscribe(fn: (event: unknown) => void): unknown;
+        },
+        streamId: string,
+      ): () => void;
+    };
+    runtimeUnavailableTools?: readonly string[];
+  },
+) => Promise<void>;
+
+export function disposeAfterTest<T extends { dispose(): void }>(value: T): T {
+  onTestFinished(() => value.dispose());
+  return value;
+}
+
+export function mockLoggerModule(
+  loggerErrorSpy?: ReturnType<typeof vi.fn>,
+): void {
+  vi.doMock('@logger', () => ({
+    createChannelTrace: () => ({
+      emit: vi.fn(),
+      debug: () => {},
+      info: () => {},
+      warn: () => {},
+      error: loggerErrorSpy ?? (() => {}),
+    }),
+  }));
+}
+
+/**
+ * Minimal AgentTrace stand-in for bridging terminal results into a session.
+ */
+export function makeFakeTrace(): {
+  subscribe(fn: (event: unknown) => void): () => void;
+  emit(event: unknown): void;
+} {
+  const subscribers = new Set<(event: unknown) => void>();
+  return {
+    subscribe(fn) {
+      subscribers.add(fn);
+      return () => subscribers.delete(fn);
+    },
+    emit(event) {
+      for (const fn of subscribers) fn(event);
+    },
+  };
+}
